Initial commit to support alias request perms.
Implemented support for global puppeteer alias permissions and possibly pair puppeteers (this is untested).
Removes some superfluous validation logic for the HandleTextAction used by Alias's and triggers to dispatch the requests.
Adds the configuration options and localization placeholders for the global option.
Fixed the regression in the Puppeteer Trigger UI to allow permissions to be set and fix a minor bug introduced in gs 1.2