Fixes alt-account authentication deletion.
Currently, as written, the code will delete any accounts not found in the connectionDto, but the connectionDto only returns the current profileUID.
If we want to mark them as stale, we could, but it is not necessary to delete them. Should leave it up to the users since it has the potential to delete the main account at present.

_Technically_ the bug is in the server GagspeakHub.cs:118, but I don't see a good way to fix this elegantly on the server side. It doesn't do any harm to leave stale user accounts on the client either (they can delete them manually as needed).

Signed-off-by: Switchy Nia <switchy-nia@users.noreply.github.com>